Fix index.d.ts declarations (#1095)
