+++ lib/authentication/auth_default.js
@@ -1,4 +1,4 @@
-const mfaAuthenticator = require ('./authentication_types.js').USER_PWD_MFA_AUTHENTICATOR;
+const mfaAuthenticator = require('./authentication_types').USER_PWD_MFA_AUTHENTICATOR;
 
 /**
  * Creates a default authenticator.
@@ -36,7 +36,7 @@ function AuthDefault(connectionConfig) {
   function setMFASessionParams(body) {
     body['data']['TOKEN'] = mfaToken;
     body['data']['AUTHENTICATOR'] = mfaAuthenticator;
-    
+
     if (isPasscodeInPassword) {
       body['data']['EXT_AUTHN_DUO_METHOD'] = 'passcode';
       body['data']['passcodeInPassword'] = true;
@@ -47,7 +47,7 @@ function AuthDefault(connectionConfig) {
       body['data']['EXT_AUTHN_DUO_METHOD'] = 'push';
     }
   }
-    
+
   this.authenticate = async function () {
     return;
   };
