--- lib/authentication/authentication_types.js
@@ -1,14 +1,0 @@
-const AuthenticationTypes =
-{
-  DEFAULT_AUTHENTICATOR: 'SNOWFLAKE', // default authenticator name
-  EXTERNAL_BROWSER_AUTHENTICATOR: 'EXTERNALBROWSER',
-  KEY_PAIR_AUTHENTICATOR: 'SNOWFLAKE_JWT',
-  OAUTH_AUTHENTICATOR: 'OAUTH',
-  USER_PWD_MFA_AUTHENTICATOR: 'USERNAME_PASSWORD_MFA',
-  ID_TOKEN_AUTHENTICATOR: 'ID_TOKEN',
-  PROGRAMMATIC_ACCESS_TOKEN: 'PROGRAMMATIC_ACCESS_TOKEN',
-  OAUTH_AUTHORIZATION_CODE: 'OAUTH_AUTHORIZATION_CODE',
-  OAUTH_CLIENT_CREDENTIALS: 'OAUTH_CLIENT_CREDENTIALS',
-};
-
-module.exports = AuthenticationTypes;
