+++ lib/errors.js
@@ -1,190 +1,12 @@
 const Util = require('./util');
 const errorMessages = require('./constants/error_messages');
+const ErrorCode = require('./error_code').default;
 
-const codes = {};
-
-// 400001
-codes.ERR_INTERNAL_ASSERT_FAILED = 400001;
-codes.ERR_UNSUPPORTED_NODE_JS_VERSION = 400002;
-
-// 401001
-codes.ERR_SF_NETWORK_COULD_NOT_CONNECT = 401001;
-codes.ERR_SF_RESPONSE_FAILURE = 401002;
-codes.ERR_SF_RESPONSE_NOT_JSON = 401003;
-codes.ERR_SF_RESPONSE_INVALID_TOKEN = 401004;
-
-// 402001
-codes.ERR_LARGE_RESULT_SET_NETWORK_COULD_NOT_CONNECT = 402001;
-codes.ERR_LARGE_RESULT_SET_RESPONSE_FAILURE = 402002;
-
-// 403001
-codes.ERR_GLOBAL_CONFIGURE_INVALID_LOG_LEVEL = 403001;
-codes.ERR_GLOBAL_CONFIGURE_INVALID_DISABLE_OCSP_CHECKS = 403002;
-codes.ERR_GLOBAL_CONFIGURE_INVALID_OCSP_MODE = 403003;
-codes.ERR_GLOBAL_CONFIGURE_INVALID_JSON_PARSER = 403004;
-codes.ERR_GLOBAL_CONFIGURE_INVALID_XML_PARSER = 403005;
-codes.ERR_GLOBAL_CONFIGURE_INVALID_KEEP_ALIVE = 403006;
-codes.ERR_GLOBAL_CONFIGURE_INVALID_CUSTOM_CREDENTIAL_MANAGER = 403007;
-codes.ERR_GLOBAL_CONFIGURE_INVALID_USE_ENV_PROXY = 403008;
-
-// 404001
-codes.ERR_CONN_CREATE_MISSING_OPTIONS = 404001;
-codes.ERR_CONN_CREATE_INVALID_OPTIONS = 404002;
-codes.ERR_CONN_CREATE_MISSING_USERNAME = 404003;
-codes.ERR_CONN_CREATE_INVALID_USERNAME = 404004;
-codes.ERR_CONN_CREATE_MISSING_PASSWORD = 404005;
-codes.ERR_CONN_CREATE_INVALID_PASSWORD = 404006;
-codes.ERR_CONN_CREATE_MISSING_ACCOUNT = 404007;
-codes.ERR_CONN_CREATE_INVALID_ACCOUNT = 404008;
-codes.ERR_CONN_CREATE_MISSING_ACCESS_URL = 404009;
-codes.ERR_CONN_CREATE_INVALID_ACCESS_URL = 404010;
-codes.ERR_CONN_CREATE_INVALID_WAREHOUSE = 404011;
-codes.ERR_CONN_CREATE_INVALID_DATABASE = 404012;
-codes.ERR_CONN_CREATE_INVALID_SCHEMA = 404013;
-codes.ERR_CONN_CREATE_INVALID_ROLE = 404014;
-codes.ERR_CONN_CREATE_MISSING_PROXY_HOST = 404015;
-codes.ERR_CONN_CREATE_INVALID_PROXY_HOST = 404016;
-codes.ERR_CONN_CREATE_MISSING_PROXY_PORT = 404017;
-codes.ERR_CONN_CREATE_INVALID_PROXY_PORT = 404018;
-codes.ERR_CONN_CREATE_INVALID_STREAM_RESULT = 404019;
-codes.ERR_CONN_CREATE_INVALID_FETCH_AS_STRING = 404020;
-codes.ERR_CONN_CREATE_INVALID_FETCH_AS_STRING_VALUES = 404021;
-codes.ERR_CONN_CREATE_INVALID_REGION = 404022;
-codes.ERR_CONN_CREATE_INVALID_KEEP_ALIVE = 404023;
-codes.ERR_CONN_CREATE_INVALID_KEEP_ALIVE_HEARTBEAT_FREQ = 404024;
-codes.ERR_CONN_CREATE_INVALID_TREAT_INTEGER_AS_BIGINT = 404025;
-codes.ERR_CONN_CREATE_INVALID_PRIVATE_KEY = 404026;
-codes.ERR_CONN_CREATE_INVALID_PRIVATE_KEY_PATH = 404027;
-codes.ERR_CONN_CREATE_INVALID_PRIVATE_KEY_PASS = 404028;
-codes.ERR_CONN_CREATE_INVALID_OAUTH_TOKEN = 404029;
-codes.ERR_CONN_CREATE_INVALID_VALIDATE_DEFAULT_PARAMETERS = 404030;
-codes.ERR_CONN_CREATE_INVALID_APPLICATION = 404031;
-codes.ERR_CONN_CREATE_MISSING_PROXY_USER = 404032;
-codes.ERR_CONN_CREATE_INVALID_PROXY_USER = 404033;
-codes.ERR_CONN_CREATE_MISSING_PROXY_PASS = 404034;
-codes.ERR_CONN_CREATE_INVALID_PROXY_PASS = 404035;
-codes.ERR_CONN_CREATE_INVALID_NO_PROXY = 404036;
-codes.ERR_CONN_CREATE_INVALID_ARRAY_BINDING_THRESHOLD = 404037;
-codes.ERR_CONN_CREATE_INVALID_GCS_USE_DOWNSCOPED_CREDENTIAL = 404038;
-codes.ERR_CONN_CREATE_INVALID_FORCE_STAGE_BIND_ERROR = 404039;
-codes.ERR_CONN_CREATE_INVALID_BROWSER_TIMEOUT = 404040;
-codes.ERR_CONN_CREATE_INVALID_DISABLED_QUERY_CONTEXT_CACHE = 404041;
-codes.ERR_CONN_CREATE_INVALID_INCLUDE_RETRY_REASON = 404042;
-codes.ERR_CONN_CREATE_INVALID_CLIENT_CONFIG_FILE = 404043;
-codes.ERR_CONN_CREATE_INVALID_RETRY_TIMEOUT = 404044;
-codes.ERR_CONN_CREATE_INVALID_ACCOUNT_REGEX = 404045;
-codes.ERR_CONN_CREATE_INVALID_REGION_REGEX = 404046;
-codes.ERR_CONN_CREATE_INVALID_DISABLE_CONSOLE_LOGIN = 404047;
-codes.ERR_CONN_CREATE_INVALID_FORCE_GCP_USE_DOWNSCOPED_CREDENTIAL = 404048;
-codes.ERR_CONN_CREATE_INVALID_CLIENT_STORE_TEMPORARY_CREDENTIAL = 404049;
-codes.ERR_CONN_CREATE_INVALID_REPRESENT_NULL_AS_STRING_NULL = 404050;
-codes.ERR_CONN_CREATE_INVALID_DISABLE_SAML_URL_CHECK = 404051;
-codes.ERR_CONN_CREATE_INVALID_CLIENT_REQUEST_MFA_TOKEN = 404052;
-codes.ERR_CONN_CREATE_MISSING_HOST = 404053;
-codes.ERR_CONN_CREATE_INVALID_HOST = 404054;
-codes.ERR_CONN_CREATE_INVALID_PASSCODE_IN_PASSWORD = 404055;
-codes.ERR_CONN_CREATE_INVALID_PASSCODE = 404056;
-codes.ERR_CONN_CREATE_MISSING_PASSWORD_AND_TOKEN = 404057;
-codes.ERR_CONN_CREATE_INVALID_OUATH_AUTHORIZATION_URL = 404058;
-codes.ERR_CONN_CREATE_INVALID_OUATH_CLIENT_ID = 404059;
-codes.ERR_CONN_CREATE_INVALID_OUATH_CLIENT_SECRET = 404060;
-codes.ERR_CONN_CREATE_INVALID_OUATH_TOKEN_REQUEST_URL = 404061;
-
-// 405001
-codes.ERR_CONN_CONNECT_INVALID_CALLBACK = 405001;
-
-// 405501
-codes.ERR_CONN_CONNECT_STATUS_CONNECTING = 405501; // sql state: 08002
-codes.ERR_CONN_CONNECT_STATUS_CONNECTED = 405502; // sql state: 08002
-codes.ERR_CONN_CONNECT_STATUS_DISCONNECTED = 405503; // sql state: 08002
-codes.ERR_CONN_CREATE_INVALID_AUTH_CONNECT = 405504;
-codes.ERR_CONN_CONNECT_INVALID_CLIENT_CONFIG = 405505;
-codes.ERR_CONN_CREATE_INVALID_AUTH_UNSUPPORTED = 405506;
-codes.ERR_CONN_CREATE_AUTH_NOT_ALLOWED = 405507;
-
-// 406001
-codes.ERR_CONN_DESTROY_INVALID_CALLBACK = 406001;
-
-// 406501
-codes.ERR_CONN_DESTROY_STATUS_PRISTINE = 406501;
-codes.ERR_CONN_DESTROY_STATUS_DISCONNECTED = 406502;
-
-// 407001
-codes.ERR_CONN_REQUEST_STATUS_PRISTINE = 407001; // sql state: 08003
-codes.ERR_CONN_REQUEST_STATUS_DISCONNECTED = 407002; // sql state: 08003
-
-// 408001
-codes.ERR_CONN_DESERIALIZE_MISSING_CONFIG = 408001;
-codes.ERR_CONN_DESERIALIZE_INVALID_CONFIG_TYPE = 408002;
-codes.ERR_CONN_DESERIALIZE_INVALID_CONFIG_FORM = 408003;
-
-// 409001
-codes.ERR_CONN_EXEC_STMT_MISSING_OPTIONS = 409001;
-codes.ERR_CONN_EXEC_STMT_INVALID_OPTIONS = 409002;
-codes.ERR_CONN_EXEC_STMT_MISSING_SQL_TEXT = 409003;
-codes.ERR_CONN_EXEC_STMT_INVALID_SQL_TEXT = 409004;
-codes.ERR_CONN_EXEC_STMT_INVALID_INTERNAL = 409005;
-codes.ERR_CONN_EXEC_STMT_INVALID_PARAMETERS = 409006;
-codes.ERR_CONN_EXEC_STMT_INVALID_BINDS = 409007;
-codes.ERR_CONN_EXEC_STMT_INVALID_BIND_VALUES = 409008;
-codes.ERR_CONN_EXEC_STMT_INVALID_COMPLETE = 409009;
-codes.ERR_CONN_EXEC_STMT_INVALID_STREAM_RESULT = 409010;
-codes.ERR_CONN_EXEC_STMT_INVALID_FETCH_AS_STRING = 409011;
-codes.ERR_CONN_EXEC_STMT_INVALID_FETCH_AS_STRING_VALUES = 409012;
-codes.ERR_CONN_EXEC_STMT_INVALID_REQUEST_ID = 409013;
-codes.ERR_CONN_EXEC_STMT_INVALID_ASYNC_EXEC = 409014;
-codes.ERR_CONN_EXEC_STMT_INVALID_DESCRIBE_ONLY = 409015;
-
-// 410001
-codes.ERR_CONN_FETCH_RESULT_MISSING_OPTIONS = 410001;
-codes.ERR_CONN_FETCH_RESULT_INVALID_OPTIONS = 410002;
-codes.ERR_CONN_FETCH_RESULT_MISSING_QUERY_ID = 410003;
-codes.ERR_CONN_FETCH_RESULT_INVALID_QUERY_ID = 410004;
-codes.ERR_CONN_FETCH_RESULT_INVALID_COMPLETE = 410005;
-codes.ERR_CONN_FETCH_RESULT_INVALID_STREAM_RESULT = 410006;
-codes.ERR_CONN_FETCH_RESULT_INVALID_FETCH_AS_STRING = 410007;
-codes.ERR_CONN_FETCH_RESULT_INVALID_FETCH_AS_STRING_VALUES = 410008;
-codes.ERR_CONN_FETCH_RESULT_INVALID_CWD = 410009;
-
-// 411001
-codes.ERR_STMT_STREAM_ROWS_INVALID_OPTIONS = 411001;
-codes.ERR_STMT_STREAM_ROWS_INVALID_START = 411002;
-codes.ERR_STMT_STREAM_ROWS_INVALID_END = 411003;
-codes.ERR_STMT_STREAM_ROWS_INVALID_FETCH_AS_STRING = 411004;
-codes.ERR_STMT_STREAM_ROWS_INVALID_FETCH_AS_STRING_VALUES = 411005;
-codes.ERR_STMT_STREAM_ROWS_INVALID_ROW_MODE = 411006;
-
-// 412001
-codes.ERR_OCSP_REVOKED = 412001;
-codes.ERR_OCSP_UNKNOWN = 412002;
-codes.ERR_OCSP_NO_SIGNATURE_ALGORITHM = 412003;
-codes.ERR_OCSP_INVALID_SIGNATURE = 412004;
-codes.ERR_OCSP_NO_RESPONSE = 412005;
-codes.ERR_OCSP_INVALID_VALIDITY = 412006;
-codes.ERR_OCSP_UNKNOWN_STATE = 412007;
-codes.ERR_OCSP_NOT_TWO_ELEMENTS = 412008;
-codes.ERR_OCSP_CACHE_EXPIRED = 412009;
-codes.ERR_OCSP_FAILED_PARSE_RESPONSE = 412010;
-codes.ERR_OCSP_INVALID_CERTIFICATE_VALIDITY = 412011;
-codes.ERR_OCSP_RESPONDER_TIMEOUT = 412012;
-codes.ERR_OCSP_CACHE_SERVER_TIMEOUT = 412013;
-codes.ERR_OCSP_FAILED_OBTAIN_OCSP_RESPONSE = 412014;
-
-// 450001
-codes.ERR_STMT_FETCH_ROWS_MISSING_OPTIONS = 450001;
-codes.ERR_STMT_FETCH_ROWS_INVALID_OPTIONS = 450002;
-codes.ERR_STMT_FETCH_ROWS_MISSING_EACH = 450003;
-codes.ERR_STMT_FETCH_ROWS_INVALID_EACH = 450004;
-codes.ERR_STMT_FETCH_ROWS_MISSING_END = 450005;
-codes.ERR_STMT_FETCH_ROWS_INVALID_END = 450006;
-codes.ERR_STMT_FETCH_ROWS_FETCHING_RESULT = 450007;
-
-// 460001
-codes.ERR_GET_RESPONSE_QUERY_INVALID_UUID = 460001;
-codes.ERR_GET_RESULTS_QUERY_ID_NO_DATA = 460002;
-codes.ERR_GET_RESULTS_QUERY_ID_NOT_SUCCESS_STATUS = 460003;
-
-exports.codes = codes;
+/**
+ * @deprecated Use `ErrorCode` instead
+ */
+exports.codes = ErrorCode;
+exports.ErrorCode = ErrorCode;
 
 /**
  * A map in which the keys are the error codes and the values are the
@@ -233,7 +55,7 @@ exports.assertInternal = function (expression, message) {
   if (!expression) {
     throw createError(errorTypes.InternalAssertError,
       {
-        code: codes.ERR_INTERNAL_ASSERT_FAILED,
+        code: ErrorCode.ERR_INTERNAL_ASSERT_FAILED,
         message: message,
         synchronous: true
       });
