+++ package.json
@@ -63,9 +63,9 @@
     "url": "https://github.com/snowflakedb/snowflake-connector-nodejs"
   },
   "main": "./dist/index.js",
-  "typings": "./index.d.ts",
+  "typings": "./dist/index.d.ts",
   "scripts": {
-    "prepack": "tsc",
+    "prepack": "node ci/build_typescript.js",
     "lint:check": "eslint && check-dts index.d.ts",
     "lint:check:all": "eslint lib samples system_test test && check-dts index.d.ts",
     "lint:check:all:errorsOnly": "npm run lint:check:all -- --quiet",
