+++ tsconfig.json
@@ -1,15 +1,4 @@
 {
-  "compilerOptions": {
-    "module": "commonjs",
-    "target": "ES2022",
-    "lib": ["ES2022"],
-    "outDir": "./dist",
-    "sourceMap": true,
-    "allowJs": true,
-    "esModuleInterop": true,
-    "resolveJsonModule": true,
-    "strict": true,
-    "skipLibCheck": true,
-  },
-  "include": ["index.js", "lib/**/*"],
+  "extends": "./tsconfig.base.json",
+  "include": ["index.js", "lib/**/*"]
 }
